Add static_tf and refactor launch files
The launch files have been moved to the kingfisher_bringup as a best
practice for better readability. The robot.launch references the other
launch files and can be used to start the relevant packages in the USV.
test.launch is also included. It does not start the rosserial to prevent
activating the motors. It is meant to be used in the lab for testing.