Update launch files and add Drive msg visualization
The launch files are updated to work with the Lidar and Camera after the
ubuntu distribution upgrade (and ros version) to Ubuntu 18.04 and ros
melodic. These changes have been used for more than a year already and
are compatible with the installation of the SBG imu.
The kingfisher_drive_viz publishes markers that are used when testing